If a package depends on React, read Symbols from its internals
This ensures that if a Symbol polyfill is loaded between React and a renderer, they no longer disagree about $$typeof values.

The shim is typechecked to match in export types to the real file. This required adding @flow to React entry point.

This increases the file size a little bit. But I think helping prevent weird load ordering issues is worth it.